xrGame: ~ исправления потенциальных вылетов из-за кривых файлов. ~ исправление пропажи гранаты из инвентаря при определенных обстоятельствах. ~ исправление пропажи звука дождя после сохранения\загрузки. ~ тестовое исправление метода set_position для инвентарных предметов (для них позиция не менялась)! ~ тестовое исправление и улучшение механизма создания иконки для драд-дропа в инвентаре (заливка и текст переносятся при драге). ~ небольшое исправление фокуса трек-бара. ~ исправление отыгрывания новый анимаций для БМ-16. + коллбек callback.on_throw_grenade на бросок гранаты игроком. Передается объект гранаты. + коллбек callback.on_weapon_shell_drop при отыгрывании партиклов гильз. Передается объект оружия, позиция гильзы, вектор скорости. + анимация anim_show2 для гранат, болта. Проигрывается после броска (если ее нету, то проигрывается анимация появления, как и было). + метод is_zoom_aiming_mode для класс CActor. Возвращает истину, когда игрок целиться из оружия. + методы SetTextColorD, SetTextColorH, SetTextColorT для класс CUI3tButton. + коллбек callback.on_cell_item_mouse отлавливает события ui_events.DRAG_DROP_ITEM_DRAG и ui_events.DRAG_DROP_ITEM_DROP.+ + метод SetColorAll для CUIStatic. Работает аналогично SetColor, но заливая при этом и детей.
О багах обязательно сообщайте в скайп inelisoni!

+++ xray/xr_3da/SH_Texture.cpp
@@ -276,7 +276,7 @@ void CTexture::LoadImpl		()
 			_fs->r_string	(buffer,sizeof(buffer));
 		}
 		double fps	= atof(buffer);
-		seqMSPF	= iFloor(1000.f/fps);
+		seqMSPF	= iFloor(1000.f/(float)fps);
 
 		while (!_fs->eof())
 		{
